% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.plot.R
\name{diag.plot}
\alias{diag.plot}
\title{Trace Plots and Density Plots for Clustered SS-PSE Model Fits}
\usage{
diag.plot(fit, type = c("trace", "density"), titl = NULL)
}
\arguments{
\item{fit}{clustered sspse fit object or a list of individual sspse fits
from each cluster.}

\item{type}{character string; type of diagnostic plot. Either \code{"trace"}
for trace plots or \code{"density"} for density plots.}

\item{titl}{character string; title of the plot.}
}
\description{
Displays MCMC trace plots and density plots for the population
size, the degree distribution parameters mu and sigma, the proportion of the
population in each cluster, and the cluster sizes. These plots show the
sampled values for each parameter at each iteration of the
Metropolis-Hastings algorithm and the overall densities produced. Plots are
displayed using the \code{ggplot2} and \code{gridExtra} packages.
}
\examples{

data(csamp)
# Make sure data is ordered by sample order
csamp <- csamp[order(csamp$samp.order), ]

# First, the Extended Bayesian model
# interval, burnin, and samplesize should be larger in a real application.
fit.eb <- posteriorsize.c(csamp$network.size.variable,
                          csamp$cluster,
                          median.prior.size = 3000,
                          prop.prior.params = rep(1, 3),
                          burnin = 100, samplesize = 500, interval = 1)

# Investigate the Extended Bayesian fit using MCMC diagnostics
diag.plot(fit.eb, type = "trace",
          titl = "Trace Plots for the Extended Bayesian Model")
diag.plot(fit.eb, type = "density",
          titl = "Density Plots for the Extended Bayesian Model")

# Next, the Posterior Sum model
csamp1 <- subset(csamp, cluster == 1)
csamp2 <- subset(csamp, cluster == 2)
csamp3 <- subset(csamp, cluster == 3)

# Obtain individual sspse fits using the sspse package
# interval, burnin, and samplesize should be larger in a real application.
library(sspse)
fit1 <- posteriorsize(csamp1$network.size.variable,
                      median.prior.size = 1500, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)
fit2 <- posteriorsize(csamp2$network.size.variable,
                      median.prior.size = 750, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)
fit3 <- posteriorsize(csamp3$network.size.variable,
                      median.prior.size = 750, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)

# Investigate the individual sspse fits using MCMC diagnostics

diag.plot(list(fit1, fit2, fit3), type = "trace",
          titl = "Trace Plots for the Posterior Sum Model")
diag.plot(list(fit1, fit2, fit3), type = "density",
          titl = "Density Plots for the Posterior Sum Model")

}
\references{
Auguie, B. (2017) \pkg{gridExtra}: Miscellaneous Functions for "Grid"
Graphics. Version 2.3. \url{https://CRAN.R-project.org/package=gridExtra}

Gamble, Laura J. (2021) \emph{Estimating the Size of Clustered Hidden
Populations}, Ph.D. Thesis, Department of Statistics, Oregon State
University.

Gile, Krista J. (2011) \emph{Improved Inference for Respondent-Driven
Sampling Data With Application to HIV Prevalence Estimation}, Journal of the
American Statistical Association, 106, 493, 135-146.

Gile, Krista J. and Handcock, Mark S. (2018) \pkg{sspse}: Estimating Hidden
Population Size using Respondent Driven Sampling Data. Los Angeles, CA.
Version 0.8, \url{http://hpmrg.org}.

Handcock, Mark S., Gile, Krista J. and Mar, Corinne M. (2014)
\emph{Estimating Hidden Population Size using Respondent-Driven Sampling
Data}, Electronic Journal of Statistics, 8, 1, 1491-1521

Wickham, H. (2016) \pkg{ggplot2}: Elegant Graphics for Data Analysis.
Springer-Verlag, New York. \url{https://ggplot2.tidyverse.org}.
}
