% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.pop.R
\name{generate.pop}
\alias{generate.pop}
\title{Generate a Clustered Population Network}
\usage{
generate.pop(N, m, probs, degs)
}
\arguments{
\item{N}{count; the overall population size}

\item{m}{count; the number of clusters in the population}

\item{probs}{vector of length \code{m}; the proportion of the population in
each cluster. Should sum to 1.}

\item{degs}{list of length \code{m}; the fully specified degree distribution
in each cluster. Should be a list of vectors, each of length \eqn{p_i N} for
cluster \eqn{i}.}
}
\value{
A list consisting of the following elements:

\item{traits}{data frame; gives the id, degree, and cluster membership of
each population member.}
\item{adj.mat}{list of length \code{m}; adjacency matrices for each cluster.}
}
\description{
Simulates a clustered population network given the population
structure and target degree distribution.
}
\details{
Networks are generated according to the configuration model. Once
members are assigned to clusters, and a degree distribution is specified,
edge ends within cluster are uniformly randomly matched to produce the
final network.
}
\examples{

N <- 500
m <- 2
probs <- c(0.75, 0.25)

d1 <- rnbinom(N*probs[1], 15, mu = 7)
d2 <- rnbinom(N*probs[2], 25, mu = 13)

cpop <- generate.pop(N, m, probs, list(d1, d2))

}
\references{
Gamble, Laura J. (2021) \emph{Estimating the Size of Clustered Hidden
Populations}, Ph.D. Thesis, Department of Statistics, Oregon State
University.

Gile, Krista J. (2011) \emph{Improved Inference for Respondent-Driven
Sampling Data With Application to HIV Prevalence Estimation}, Journal of the
American Statistical Association, 106, 493, 135-146.
}
