\name{netData}
\alias{netData}
\docType{data}
\title{Dataset: netData}
\description{Network data with 10 nodes and 2 layers}
\usage{data("netData")}
\format{
  The format is:
 num [1:10, 1:10, 1:2, 1:2] 0 0 0 0 0 0 0 0 0 0 ...
}
\details{Dataset demonstrating multilayer network}
\source{Sarkar, S. (2020)}
\references{
Sarkar, S. (2019) On the use of transformations for modeling multidimensional heterogeneous data, The University of Alabama Libraries Digital Collections
}
\examples{
data(netData)
## maybe str(netData) ; plot(netData) ...
}
\keyword{datasets}
