\name{netDataID}
\alias{netDataID}
\docType{data}
\title{Dataset: netDataID}
\description{ID for netData dataset}
\usage{data("netDataID")}
\format{
  A data frame with 10 observations on the following 1 variable.
  \describe{
    \item{\code{netDataID}}{a numeric vector}
  }
}
\details{ID for the dataset demonstrating multilayer network}
\source{Sarkar, S. (2020)}
\references{
Sarkar, S. (2019) On the use of transformations for modeling multidimensional heterogeneous data, The University of Alabama Libraries Digital Collections
}
\examples{
data(netDataID)
## maybe str(netDataID) ; plot(netDataID) ...
}
\keyword{datasets}
