% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.plot.R
\name{post.plot}
\alias{post.plot}
\title{Plot the Posterior Distributions Given by the Extended Bayesian
and Posterior Sum Models}
\usage{
post.plot(
  sspse.obj,
  type = c("N", "p", "N.c"),
  band.post = 1,
  n = NULL,
  zoom = 0.975,
  nrow = NULL,
  titl = "Posterior for Population Size"
)
}
\arguments{
\item{sspse.obj}{when using the Extended Bayesian model, a clustered sspse
fit object for plot types \code{N}, \code{p}, or \code{N.c}. When using the
Posterior Sum model, a vector of draws from the posterior sum distribution
for plot type \code{N} and a list of individual sspse fit objects for plot
type \code{N.c}.}

\item{type}{character string; type of posterior plot. Either \code{"N"} for
overall population size, \code{"p"} for population proportions, or
\code{"N.c"} for cluster sizes.}

\item{band.post}{scalar; multiplication factor for the smoothing bandwidth
used in kernel density estimation.}

\item{n}{count; sample size of the RDS sample. This can be extracted from
sspse fit objects.}

\item{zoom}{upper x-axis limit given as a quantile of the posterior
distribution.}

\item{nrow}{count; number of rows passed to \code{\link[ggplot2]{facet_wrap}}.}

\item{titl}{character string; title of the plot.}
}
\description{
Plots the posterior distributions for overall population size,
\eqn{P(N | Data)}; for the marginal population proportions, \eqn{P(p_i | Data)}
in each cluster \eqn{i}; and for the marginal cluster sizes,
\eqn{P(N_i | Data)} in each cluster \eqn{i}.
}
\examples{

data(csamp)
# Make sure data is ordered by sample order
csamp <- csamp[order(csamp$samp.order), ]

# First, the Extended Bayesian model
# interval, burnin, and samplesize should be larger in a real application.
fit.eb <- posteriorsize.c(csamp$network.size.variable,
                          csamp$cluster,
                          median.prior.size = 3000,
                          prop.prior.params = rep(1, 3),
                          burnin = 100, samplesize = 500, interval = 1)

post.plot(fit.eb, type = "N",
          titl = "Posterior for Overall Size: Extended Bayesian")
post.plot(fit.eb, type = "p")
post.plot(fit.eb, type = "N.c", zoom = 1, nrow = 3)

# Next, the Posterior Sum model
csamp1 <- subset(csamp, cluster == 1)
csamp2 <- subset(csamp, cluster == 2)
csamp3 <- subset(csamp, cluster == 3)

# Obtain individual sspse fits using the sspse package
# interval, burnin, and samplesize should be larger in a real application.
library(sspse)
fit1 <- posteriorsize(csamp1$network.size.variable,
                      median.prior.size = 1500, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)
fit2 <- posteriorsize(csamp2$network.size.variable,
                      median.prior.size = 750, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)
fit3 <- posteriorsize(csamp3$network.size.variable,
                      median.prior.size = 750, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)

# Use the Posterior Sum model to combine the individual fits
fit.ps <- post.sum(list(fit1, fit2, fit3),
                   samplesize = 5000)

post.plot(fit.ps, type = "N", n = 1500, zoom = 1)
post.plot(list(fit1, fit2, fit3), type = "N.c", zoom = 1, nrow = 3)

}
\references{
Gamble, Laura J. (2021) \emph{Estimating the Size of Clustered Hidden
Populations}, Ph.D. Thesis, Department of Statistics, Oregon State
University.

Gile, Krista J. (2011) \emph{Improved Inference for Respondent-Driven
Sampling Data With Application to HIV Prevalence Estimation}, Journal of the
American Statistical Association, 106, 493, 135-146.

Gile, Krista J. and Handcock, Mark S. (2018) \pkg{sspse}: Estimating Hidden
Population Size using Respondent Driven Sampling Data. Los Angeles, CA.
Version 0.8, \url{http://hpmrg.org}.

Handcock, Mark S., Gile, Krista J. and Mar, Corinne M. (2014)
\emph{Estimating Hidden Population Size using Respondent-Driven Sampling
Data}, Electronic Journal of Statistics, 8, 1, 1491-1521

Wickham, H. (2016) \pkg{ggplot2}: Elegant Graphics for Data Analysis.
Springer-Verlag, New York. \url{https://ggplot2.tidyverse.org}.
}
