% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.RDS.c.R
\name{sample.RDS.c}
\alias{sample.RDS.c}
\title{Draw an RDS Sample from a Clustered Population}
\usage{
sample.RDS.c(
  traits,
  adj.mat,
  N.c,
  id = "id",
  cluster = "cluster",
  n,
  num.seeds,
  num.samp,
  num.prob,
  replace = FALSE
)
}
\arguments{
\item{traits}{data frame; including unique ID numbers and cluster membership
for all nodes in the population.}

\item{adj.mat}{list of length \code{m} or matrix; each entry being an
adjacency matrix from each cluster. Alternately a single adjacency matrix
for the whole population. If a single adjacency matrix is supplied,
\code{N.c} must be supplied as well.}

\item{N.c}{vector of length \code{m}; the population size in each cluster.}

\item{id}{character string; the column name in \code{traits} for the unique ID
number of each member. These should be the integers 1, ..., \code{sum(N.c)}
that directly correspond to the order of the adjacency matrix.}

\item{cluster}{character string; the column name in \code{traits} for the cluster
membership of each member.}

\item{n}{vector of length \code{m} or count; the target sample sizes to be
drawn from each cluster. If a single value is entered, that sample size will
be used for all clusters.}

\item{num.seeds}{vector of length \code{m} or count; the number of seeds in
each cluster. If a single value is entered, that number of seeds will be
used for all clusters.}

\item{num.samp}{vector of length \code{m} or count; the maximum number of
recruits for a respondent in each cluster. If a single value is entered,
that number of possible recruits will be used for all clusters.}

\item{num.prob}{list of length \code{m} or vector of length
\code{num.samp} + 1; each entry being a vector of positive numbers with
length \code{num.samp} + 1. Each vector represents the probability of a
respondent in the corresponding cluster recruiting 0, 1, ..., \code{num.samp}
new respondents. If a single vector is entered, that vector of probabilities
will be used for all clusters.}

\item{replace}{Boolean; whether sampling is performed with replacement.}
}
\value{
A data frame with columns

\item{id}{each respondent's unique ID number.}
\item{order}{the order in which each respondent was sampled within their
cluster.}
\item{recruiter.id}{unique ID number for the recruiter of each respondent.}
\item{network.size.variable}{degree of each respondent.}
\item{cluster}{true cluster membership.}
\item{seed.cluster}{cluster membership of the corresponding seed for each
respondent.}
}
\description{
Simulates an RDS sample from each cluster in a clustered
population with \code{m} clusters. This method of approximating an RDS
process is a clustered extension of the
\code{\link[RDStreeboot]{sample.RDS}} function in the \code{RDStreeboot}
package.
}
\details{
This function simulates an RDS sample from a given clustered social
network. Seeds are drawn with probability proportional to unit size within
their clusters. Each respondent is recruited instantly and in order according
to the recruitment probabilities supplied. Recruitment is done uniformly
randomly.

The network can be supplied as a list of adjacency matrices in each cluster
or as an overall adjacency matrix of the entire population. Networks can
therefore be supplied in which connections between clusters exist. In this
case, sampling will allow for travel between clusters. The assumed cluster
membership, or the membership of each respondent's seed, is reported along
with the true cluster membership for each individual in the sample.
}
\examples{

# First, generate a clustered population
N <- 500
m <- 2
probs <- c(0.75, 0.25)

d1 <- rnbinom(N*probs[1], 15, mu = 7)
d2 <- rnbinom(N*probs[2], 25, mu = 13)

cpop <- generate.pop(N, m, probs, list(d1, d2))

RDS.csamp <- sample.RDS.c(cpop[[1]], cpop[[2]],
                          id = "node", n = c(75, 50),
                          num.seeds = 2, num.samp = 3,
                          num.prob = c(0.1, 0.4, 0.4, 0.1))

}
\references{
Baraff, Aaron J. (2016). \pkg{RDStreeboot}: RDS Tree Bootstrap Method.
Version 1.0. \url{https://CRAN.R-project.org/package=RDStreeboot}.

Gamble, Laura J. (2021) \emph{Estimating the Size of Clustered Hidden
Populations}, Ph.D. Thesis, Department of Statistics, Oregon State
University.

Gile, Krista J. (2011) \emph{Improved Inference for Respondent-Driven
Sampling Data With Application to HIV Prevalence Estimation}, Journal of the
American Statistical Association, 106, 493, 135-146.

Handcock, Mark S., Fellows, Ian E., and Gile, Krista J. (2012) \pkg{RDS}:
Respondent-Driven Sampling. Los Angeles, CA. Version 0.9-2,
\url{http://hpmrg.org}.
}
