% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.c.R
\name{summary.c}
\alias{summary.c}
\title{Summary Statistics for Clustered SS-PSE Model Fits}
\usage{
summary.c(fit, support = 512, HPD.level = 0.95)
}
\arguments{
\item{fit}{clustered sspse fit object or a list of individual sspse fits
from each cluster or a list of individual sspse fits from each cluster and
a sample from the posterior sum distribution.}

\item{support}{count; the number of equally spaced points at which the posterior
densities are estimated.}

\item{HPD.level}{scalar; probability level of the highest probability
density interval determined from the estimated posterior.}
}
\description{
Prints summary information about the prior and posterior
distributions of a given clustered sspse fit.
}
\examples{

data(csamp)
# Make sure data is ordered by sample order
csamp <- csamp[order(csamp$samp.order), ]

# First, the Extended Bayesian model
# interval, burnin, and samplesize should be larger in a real application.
fit.eb <- posteriorsize.c(csamp$network.size.variable,
                          csamp$cluster,
                          median.prior.size = 3000,
                          prop.prior.params = rep(1, 3),
                          burnin = 100, samplesize = 500, interval = 1)

summary.c(fit.eb)

# Next, the Posterior Sum model
csamp1 <- subset(csamp, cluster == 1)
csamp2 <- subset(csamp, cluster == 2)
csamp3 <- subset(csamp, cluster == 3)

# Obtain individual sspse fits using the sspse package
# interval, burnin, and samplesize should be larger in a real application.
library(sspse)
fit1 <- posteriorsize(csamp1$network.size.variable,
                      median.prior.size = 1500, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)
fit2 <- posteriorsize(csamp2$network.size.variable,
                      median.prior.size = 750, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)
fit3 <- posteriorsize(csamp3$network.size.variable,
                      median.prior.size = 750, K = 100,
                      burnin = 100, samplesize = 500, interval = 1)

# Use the Posterior Sum model to combine the individual fits
fit.ps <- post.sum(list(fit1, fit2, fit3),
                   samplesize = 5000)

summary.c(list(fit1, fit2, fit3, fit.ps),
          HPD.level = 0.9, support = 10000)

}
\references{
Gamble, Laura J. (2021) \emph{Estimating the Size of Clustered Hidden
Populations}, Ph.D. Thesis, Department of Statistics, Oregon State
University.

Gile, Krista J. (2011) \emph{Improved Inference for Respondent-Driven
Sampling Data With Application to HIV Prevalence Estimation}, Journal of the
American Statistical Association, 106, 493, 135-146.

Gile, Krista J. and Handcock, Mark S. (2018) \pkg{sspse}: Estimating Hidden
Population Size using Respondent Driven Sampling Data. Los Angeles, CA.
Version 0.8, \url{http://hpmrg.org}.

Handcock, Mark S., Gile, Krista J. and Mar, Corinne M. (2014)
\emph{Estimating Hidden Population Size using Respondent-Driven Sampling
Data}, Electronic Journal of Statistics, 8, 1, 1491-1521
}
