% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.concentration.R
\name{visualize.concentration}
\alias{visualize.concentration}
\title{Plot the Marginal Distributions of a Dirichlet Random Variable Under
Different Concentration Parameters}
\usage{
visualize.concentration(
  conc.values,
  p,
  nrow = NULL,
  titl = NULL,
  low.col = "#E7AD2B",
  high.col = "darkblue"
)
}
\arguments{
\item{conc.values}{vector; different values for the concentration parameter
of a Dirichlet distribution.}

\item{p}{vector; the mean vector of the Dirichlet distribution.}

\item{nrow}{count; number of rows passed to \code{\link[ggplot2]{facet_wrap}}.}

\item{titl}{character string; title of the plot.}
}
\description{
Plots the marginal distributions of a Dirichlet random variable.
Useful for visualizing the effect of the concentration parameter on a
Dirichlet distribution with a set mean vector \code{p}.
}
\examples{

# Visualize the effect of concentration parameters from 1 to 10 on a
# Dirichlet distribution with mean vector (0.5, 0.25, 0.25)
visualize.concentration(1:10, c(0.5, 0.25, 0.25))
visualize.concentration(1:10, c(0.5, 0.25, 0.25), nrow = 3)

}
\references{
Gamble, Laura J. (2021) \emph{Estimating the Size of Clustered Hidden
Populations}, Ph.D. Thesis, Department of Statistics, Oregon State
University.

Wickham, H. (2016) \pkg{ggplot2}: Elegant Graphics for Data Analysis.
Springer-Verlag, New York. \url{https://ggplot2.tidyverse.org}.
}
